lisp: bugfix: finally fixed (unsigned-byte 8) bug
The function which gets called by usocket:socket-server when a message
is received will be checked for a return value. If a return value is
present it will be sent. Since this was not documented it threw a error
when the value was not a sequence of (unsigned-byte 8). fixed! :D