lisp: :directory-deleted now working if RECURSIVE-P NIL
The event :directory-deleted was not working since if RECURSIVE-P was
NIL no subdirectory was tracked, so it was not possible to distinguish
between :directory-delted and :file-deleted. To Fix this issue all
subdirectories (not recursivly) below the main directory are now getting
tracked too, just without a handle attached to them.