lisp: fixed path beeing the absolute path
The pathname which is transfered to other clients should not be the
absolute path. Instead it should be a relative path starting with the
shared directory under Root. For example, if i want to share the folders
/home/d4ryus/share-me/ and /home/share-me-too/ it should display these
two folders like: /share-me/ and /share-me-too/ not showing where they
are on the local system.