lisp: moved LIST-OF-CHANGES to lodds-server
since its possible to have multiple watchers its more fitting to have
the list of changes in one place and push events to it. So with this
change i also implemented a CHANGE-HOOK which will be called by the
watcher if anything changes, this hook will then push the change onto
LIST-OF-CHANGES inside the lodds-server. Also moved all corresponding
functions from watcher.lisp to lodds.lisp, since they are accessing
the lodds-server object now instead of watcher object.