lisp: implemented STOP-WATCHER
If a user has multiple folders shared and removes one of them, the files
inside that folder need to be removed from the shared list. STOP-WATCHER
will add a :del change for every file tracked by the removed
watcher. Its also possible to prevent this behavior by calling
STOP-WATCHER is run-change-hook-p set to nil. Thats handy if u want to
stop all watcher and remove the list, because then there is not need to
add all those change events.