lisp: bugfix: added CURRENT-TIMESTAMP to GET-FILE-CHANGES
we need the CURRENT-TIMESTAMP to make a clean cut between what changes
will be added to the list and which not. So if a client requests a
change with a timestamp he gets all changes from
CURRENT-TIMESTAMP (excluding CURRENT-TIMESTAMP) until
timestamp (including timestamp). He will then get CURRENT-TIMESTAMP as
'revision'.