lisp: implemented new subsystems, started working on event-queue
this is a way more generic way to handle all kinds of subsystems. Each
is now a instance of the SUBSYSTEM class, which contains a name, thread,
alive-p, init-fn, inti-args and a event-queue. Subsystems can now be
started and stopped with SUBSYSTEM-START and SUBSYSTEM-STOP. Error
handling is now in one place too, which is nice.

next up:
  - update watcher to be a subsystem too
  - hook into the event-queue