lisp: made the lodds object a global variable
This will extremly simplify the code since there is no need to pass
around the server or any subsystem. The current server object is just
simply bound to lodds:*server* now. Also added a simple
macro (WITH-SERVER) to dynamically bind a server and add the binding
to bt:*DEFAULT-SPECIAL-BINDINGS* to get the wanted binding inside all
spawned threads. This will ensure that all spawned thread act on the
bound server.

Also replaced all Methods which where not 'needed' with normal
Functions.