lisp: removes STMX as dependancy
Using STMX solved the Problem, but it made the code way harder to read
instead of just using Locks. I guess thats because the need for
synchronisation in lodds is not to 'crazy', there are just a few
variables which will be accesses by multiple threads. And these
accesses happen in one place, so a lock is perfectly sufficent.