lisp: bugfix: fixes error if file has vanished
its possible that a :file-created event is triggered (watcher) and
if we try to open the file to calculate the sha1 the file is already
gone. This will just wrap the file open with a HANDLER-CASE and
returns a sha of 0000... This should be fine since a delete will
follow anyway. We just run into problems if we dont add the file to
the list.