lisp: bugfix: fixes bug of zombie info widgets
When we add tasks to the gui we get by get-task-progresses, we might
actually add tasks that have already been removed. This adds a check
which removes all tasks we track, but are not inside the
list (basically we are tracking a task which does not exist anymore,
since it does not exist we dont get a failed/canceld/finished event).