lisp: add setting :show-log-widget
add setting to show the log widget tabbed with the info widget at the
bottom of the main view. On some plattforms i was not able to display
the hidden log widget, this is a workaround to be able to show the log
widget on those by setting :show-log-widget to t in a config file and
restarting lodds.