lisp: bugfix: handle directory renames (mv) event
when a directory with one or more subdirectories is renamed (moved
with mv) there wont be a :directory-removed event for the
subdirectories. That behavior is the same on windows (where windows
does never throw events for subdirectories) so we can just remove the
os-windows check to always remove subdirectories by hand. This will
add a bit of overhead on non os-windows systems, but should not be
noticeable.