lisp: bugfix: add own ensure-directory-pathname
when a non directory pathname with escaped wildcards is passed to
uiop:ensure-directory-pathname it will use make-pathname to construct
a new directory pathname. This however will lead to another level of
escape characters if the :name part of the pathname is added to the
:directory part. lodds.core:ensure-directory-pathname will add a
trailing character to ensure the pathname is parsed as a directory
pathname to not add another level of escape characters.